requirements.txt
asgiref==3.4.1
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
cipher==0.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
defusedxml==0.7.1
Django==3.2.6
django-cors-headers==3.8.0
django-environ==0.4.5
django-filter==21.1
djangorestframework==3.12.4
djoser==2.1.0
drf-yasg==1.20.0
gunicorn==20.1.0
idna==3.2
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
oauthlib==3.1.1
packaging==21.0
psycopg2==2.9.1
pycparser==2.20
pycryptodome==3.10.1
PyJWT==2.1.0
pyparsing==2.4.7
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.6
Pillow==8.4.0